+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/SignalingDisconnectResult.kt
@@ -1,0 +1,6 @@
+package jp.shiguredo.sora.sdk.channel
+
+data class SignalingDisconnectResult(
+    val code: Int,
+    val reason: String,
+)