+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/data/SoraCloseEvent.kt
@@ -1,0 +1,9 @@
+package jp.shiguredo.sora.sdk.channel.data
+
+import jp.shiguredo.sora.sdk.error.SoraDisconnectReason
+
+data class SoraCloseEvent(
+    val title: SoraDisconnectReason,
+    val code: Int?,
+    val reason: String?,
+)